interpolation search

+++ search_algorithms/102-interpolation.c
@@ -21,9 +21,9 @@ int interpolation_search(int *array, size_t size, int value)
 	{
 		pos = low + (((double)(high - low) / (array[high] - array[low]))
 					* (value - array[low]));
-		if (pos > array[high] || pos < array[low])
+		if (pos > (size_t)array[high] || pos < (size_t)array[low])
 		{
-			printf("Value checked array[%d] is out of range\n", pos);
+			printf("Value checked array[%lu] is out of range\n", pos);
 			return (-1);
 		}
 		printf("Value checked array[%lu] = [%d]\n", pos, array[pos]);
